% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{cluster_distances}
\alias{cluster_distances}
\title{Clustering dissimilarities}
\usage{
cluster_distances(dissimilarity_matrix)
}
\arguments{
\item{distance_matrix}{A dissimilarity matrix (required)}
}
\value{
an hclust object
}
\description{
Clustering dissimilarities
}

