% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{convert_to_distance_matrix}
\alias{convert_to_distance_matrix}
\title{Convert pairwise dissimilarity data.frame to a pairwise matrix}
\usage{
convert_to_distance_matrix(master_diff, measure = "manhattan_dist_scaled",
  diag = NA, sample_subset = NULL)
}
\arguments{
\item{master_diff}{The data.frame containing pairwise dissimilarites. Typically the output of \code{create_pairwise_master}. Required.}

\item{measure}{The name of the column to use as the values for the pairwise dissimilarity matrix. Defaults to the manhanttan distance. Can be changed if user wants another measure (e.g. correlation, etc).}

\item{diag}{The value to put in the diagonal. Defaults to NA.}

\item{sample_subset}{A character vector of sample names to subset. Optional.}
}
\value{
A data.matrix representing the pairwise dissimilarites between every cell
}
\description{
Convert pairwise dissimilarity data.frame to a pairwise matrix
}

