% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{divide_cores}
\alias{divide_cores}
\title{Divide cores}
\usage{
divide_cores(total, ncores = cores_to_use)
}
\arguments{
\item{total}{Total number of jobs to parallelize}

\item{ncores}{CPUs to parallelize over (must be >1)}
}
\value{
Returns a data.frame with two columns (from and to) and rows equal to the number of cores
}
\description{
Utility function to help parallelize jobs in R with less overhead.
    Likely wouldn't ever be used by the user.
}
\examples{
divide_cores(total = 100, ncores = 2)

}

