% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{get_diff_df}
\alias{get_diff_df}
\title{Get Dissimilarity from a pairwise common data frame}
\usage{
get_diff_df(df)
}
\arguments{
\item{df}{A dataframe containing the pairwise common CpGs. Required.}
}
\value{
A 1-row dataframe containing the pairwise dissimilarity
}
\description{
Get Dissimilarity from a pairwise common data frame
}
\details{
Typically this function is not used by the user. It is called by \code{create_pairwise_master} when \code{calc_diff} is TRUE.

The input dataframe requires four columns in the specific order:
1. Chromosome
2. Start
3. Name of the 1st comparitor (e.g. cell1)
4. Name of the 2nd comparitor (e.g. cell2)
}

