% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{merge_cpgs}
\alias{merge_cpgs}
\title{Create in silico merged bulk profiles from single-cell files}
\usage{
merge_cpgs(cluster_members, cpg_all)
}
\arguments{
\item{cluster_members}{A vector of samples to include}

\item{cpg_all}{The list containing all CpG calls in data.frame format}
}
\value{
a BSmooth object ready for smoothing
}
\description{
Create in silico merged bulk profiles from single-cell files
}
\details{
Uses the bsseq package to perform in silico merging of single-cell CpG calls. Requires the bsseq R package to be installed
}

