% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize clusters}
\usage{
visualize_clusters(dissimilarity_matrix, cluster_results = NA)
}
\arguments{
\item{dissimilarity_matrix}{A matrix outputted by `convert_to_dissimilarity_matrix()`}

\item{cluster_results}{The result from `cluster_dissimilarity()`. If left empty, no clustering results will be visualized.}
}
\value{
A ggplot-ready data.frame
}
\description{
Visualize clusters
}

