% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMeth_functions.R
\name{merge_cpgs}
\alias{merge_cpgs}
\title{Create in silico merged bulk profiles from single-cell files}
\usage{
merge_cpgs(cpg_all, cluster_assignments, desired_cluster)
}
\arguments{
\item{cpg_all}{The list containing all CpG calls in data.frame format. Required.}

\item{cluster_assignments}{The cluster_assignments outputed from cluster_dissimilarity(). Required}

\item{desired_cluster}{The number desired from the input cluster assignments. Required}
}
\value{
a BSmooth object ready for smoothing
}
\description{
Create in silico merged bulk profiles from single-cell files
}
\details{
Uses the bsseq package to perform in silico merging of single-cell CpG calls. Requires the bsseq R package to be installed
}

